// repos/module-data-bi-report/config/types.ts
export enum EnumCategory {
  dimension = 'dimension',
  metric = 'metric'
}

export enum EnumDisplayType {
  normal = 'normal',
  date = 'date',
  int = 'int',
  float = 'float',
  intPercent = 'intPercent',
  floatPercent = 'floatPercent',
}

export class BiMetadataDto {
  fieldName: FieldName
  displayName: string
  dataType: string
  logicCategory: EnumCategory
  displayType: EnumDisplayType
  type: number
  displayGroup: string
}

export enum FieldName {
  /// type1 
  actionDay = 'action_day',
  actionHour = 'action_hour',
  agentName = 'agent_name',
  customerName = 'customer_name',
  projectName = 'project_name',
  campaignName = 'campaign_name',
  contractName = 'contract_name',
  orderName = 'order_name',
  creativeName = 'creative_name',
  frequency = 'frequency',
  adxId = 'adx_id',
  adxName = 'adx_name',
  requestTypeName = 'request_type_name',
  requestTerminal = 'request_terminal',
  displayForm = 'display_form',
  relevantType = 'relevant_type',
  dealId = 'deal_id',
  appId = 'app_id',
  appPackage = 'app_package',
  appName = 'app_name',
  appCategoryName = 'app_category_name',
  country = 'country',
  province = 'province',
  city = 'city',
  cityLevel = 'city_level',
  county = 'county',
  agentId = 'agent_id',
  customerId = 'customer_id',
  projectId = 'project_id',
  contractId = 'contract_id',
  campaignId = 'campaign_id',
  orderId = 'order_id',
  creativeId = 'creative_id',
  adId = 'ad_id',
  creativeDynamicType = 'creative_dynamic_type',
  otvPosition = 'otv_position',
  adzoneId = 'adzone_id',
  adzoneName = 'adzone_name',
  adzoneSize = 'adzone_size',
  requestMaxDuration = 'request_max_duration',
  requestMinDuration = 'request_min_duration',
  creativeVideoDuration = 'creative_video_duration',
  nativeId = 'native_id',
  nativeName = 'native_name',
  bidNativeId = 'bid_native_id',
  bidNativeName = 'bid_native_name',
  os = 'os',
  channelName = 'channel_name',
  deviceBrand = 'device_brand',
  conversionName = 'conversion_name',
  conversionOrderId = 'conversion_order_id',
  conversionDetail = 'conversion_detail',
  conversionIdentity = 'conversion_identity',
  impressionTime = 'impression_time',
  clickTime = 'click_time',
  conversionTime = 'conversion_time',
  adzonePosition = 'adzone_position',
  mainDomain = 'main_domain',
  effectDelayDay = 'effect_delay_day',
  purchaseFromType = 'purchase_from_type',
  adzoneReservePrice = 'adzone_reserve_price',
  ask = 'ask',
  bid = 'bid',
  win = 'win',
  impression = 'impression',
  click = 'click',
  reach = 'reach',
  conversion = 'conversion',
  impConversion = 'imp_conversion',
  clickConversion = 'click_conversion',
  askUv = 'ask_uv',
  bidUv = 'bid_uv',
  winUv = 'win_uv',
  impressionUv = 'impression_uv',
  clickUv = 'click_uv',
  reachUv = 'reach_uv',
  conversionUv = 'conversion_uv',
  btr = 'btr',
  wtr = 'wtr',
  ctr = 'ctr',
  ecpm = 'ecpm',
  agentEcpm = 'agent_ecpm',
  customerEcpm = 'customer_ecpm',
  cpc = 'cpc',
  agentCpc = 'agent_cpc',
  customerCpc = 'customer_cpc',
  cost = 'cost',
  agentCost = 'agent_cost',
  customerCost = 'customer_cost',
  contractCost = 'contract_cost',
  lostCost = 'lost_cost',
  conversionPrice = 'conversion_price',
  deepLinkAccess = 'deep_link_access',
  deepLinkEvoke = 'deep_link_evoke',
  deepLinkEvokeSuccess = 'deep_link_evoke_success',
  deepLinkEvokeFail = 'deep_link_evoke_fail',
}

export const displayNameMap: Map<FieldName, string> = new Map([
  [FieldName.actionDay, '日期'],
  [FieldName.actionHour, '小时'],
  [FieldName.agentName, '代理商名称'],
  [FieldName.customerName, '广告主名称'],
  [FieldName.projectName, '项目名称'],
  [FieldName.campaignName, '计划名称'],
  [FieldName.contractName, '合同名称'],
  [FieldName.orderName, '订单名称'],
  [FieldName.creativeName, '创意名称'],
  [FieldName.frequency, '频次'],
  [FieldName.adxId, '平台ID'],
  [FieldName.adxName, '平台名称'],
  [FieldName.requestTypeName, '流量采买类型'],
  [FieldName.requestTerminal, '流量终端'],
  [FieldName.displayForm, '展现形式'],
  [FieldName.relevantType, '是否原生'],
  [FieldName.dealId, 'Deal ID'],
  [FieldName.appId, 'APP ID'],
  [FieldName.appPackage, 'APP包名'],
  [FieldName.appName, 'APP名称'],
  [FieldName.appCategoryName, '媒体类型'],
  [FieldName.country, '国家'],
  [FieldName.province, '省份'],
  [FieldName.city, '城市'],
  [FieldName.cityLevel, '城市等级'],
  [FieldName.county, '区县'],
  [FieldName.agentId, '代理商ID'],
  [FieldName.customerId, '广告主ID'],
  [FieldName.projectId, '项目ID'],
  [FieldName.contractId, '合同ID'],
  [FieldName.campaignId, '计划ID'],
  [FieldName.orderId, '订单ID'],
  [FieldName.creativeId, '创意ID'],
  [FieldName.adId, '广告ID'],
  [FieldName.creativeDynamicType, '创意展现类型'],
  [FieldName.otvPosition, '贴片位置'],
  [FieldName.adzoneId, '广告位ID'],
  [FieldName.adzoneName, '广告位名称'],
  [FieldName.adzoneSize, '广告尺寸'],
  [FieldName.requestMaxDuration, '可投创意最大时长'],
  [FieldName.requestMinDuration, '可投创意最小时长'],
  [FieldName.creativeVideoDuration, '创意时长'],
  [FieldName.nativeId, '可投模板ID'],
  [FieldName.nativeName, '可投模板名称'],
  [FieldName.bidNativeId, '竞得模板ID'],
  [FieldName.bidNativeName, '竞得模板名称'],
  [FieldName.os, '操作系统'],
  [FieldName.channelName, '视频频道名称'],
  [FieldName.deviceBrand, '手机品牌'],
  [FieldName.conversionName, '转化名称'],
  [FieldName.conversionOrderId, '转化订单编号'],
  [FieldName.conversionDetail, '转化详情'],
  [FieldName.conversionIdentity, '转化用户'],
  [FieldName.impressionTime, '转化-曝光时间'],
  [FieldName.clickTime, '转化-点击时间'],
  [FieldName.conversionTime, '转化时间'],
  [FieldName.adzonePosition, '广告位置'],
  [FieldName.mainDomain, '域名'],
  [FieldName.effectDelayDay, '是否延迟'],
  [FieldName.purchaseFromType, 'WAC'],
  [FieldName.adzoneReservePrice, '广告位低价(分)'],
  [FieldName.ask, '询价量'],
  [FieldName.bid, '竞价量'],
  [FieldName.win, '竞价成功量'],
  [FieldName.impression, '曝光量'],
  [FieldName.click, '点击量'],
  [FieldName.reach, '到达量'],
  [FieldName.conversion, '总转化量'],
  [FieldName.impConversion, '曝光转化量'],
  [FieldName.clickConversion, '点击转化量'],
  [FieldName.askUv, '询价访客'],
  [FieldName.bidUv, '竞价访客'],
  [FieldName.winUv, '竞价成功访客'],
  [FieldName.impressionUv, '曝光访客'],
  [FieldName.clickUv, '点击访客'],
  [FieldName.reachUv, '到达访客'],
  [FieldName.conversionUv, '总转化访客'],
  [FieldName.btr, '竞价率'],
  [FieldName.wtr, '竞价成功率'],
  [FieldName.ctr, '点击率'],
  [FieldName.ecpm, 'eCPM(实际花费)'],
  [FieldName.agentEcpm, 'eCPM(代理商花费)'],
  [FieldName.customerEcpm, 'eCPM(广告主花费)'],
  [FieldName.cpc, 'CPC(实际花费)'],
  [FieldName.agentCpc, 'CPC(代理商花费)'],
  [FieldName.customerCpc, 'CPC(广告主花费)'],
  [FieldName.cost, '实际花费'],
  [FieldName.agentCost, '代理商花费'],
  [FieldName.customerCost, '广告主花费'],
  [FieldName.contractCost, '合同号花费'],
  [FieldName.lostCost, '花超花费'],
  [FieldName.conversionPrice, '转化金额'],
  [FieldName.deepLinkAccess, 'Deeplink中间页到达量'],
  [FieldName.deepLinkEvoke, 'Deeplink中间页起吊量'],
  [FieldName.deepLinkEvokeSuccess, 'Deeplink中间页起吊成功量'],
  [FieldName.deepLinkEvokeFail, 'Deeplink中间页起吊失败量' ],
])
export enum QueryType {
  entityQuery = 'entityQuery'
}

export enum QueryLogicType {
  bi = 'bi',
}